+++ examples/todomvc/event.rs
@@ -1,41 +1,7 @@
-use crate::{resource::*, ui_state::UiTodo};
+use crate::ui_state::UiTodo;
 use bevy_dioxus::desktop::prelude::*;
 
-// UI -> ECS
-
-#[derive(Clone, Debug)]
-pub struct CreateTodo {
-    pub title: String,
-}
-
-#[derive(Clone, Debug)]
-pub struct ChangeTitle {
-    pub entity: Entity,
-    pub title: String,
-}
-
-#[derive(Clone, Debug)]
-pub struct ToggleDone {
-    pub entity: Entity,
-}
-
-#[derive(Clone, Debug)]
-pub struct RemoveTodo {
-    pub entity: Entity,
-}
-
-#[derive(Clone, Debug)]
-pub struct ToggleAll;
-
-#[derive(Clone, Debug)]
-pub struct ChangeFilter {
-    pub filter: Filter,
-}
-
-#[derive(Clone, Debug)]
-pub struct ClearCompleted;
-
-// Internal (System -> System)
+// Internal events (System -> System)
 
 pub struct UpdateTodoMeta {
     pub entity: Entity,
